Move parse-env.ts to src/lib/
- Relocate parseEnvString utility from components/projects/ to lib/
- Update import in SecretsPanel.tsx to use @/lib/parse-env
